requirements.txt
autopep8==1.4.4
Django==2.2.2
django-appconf==1.0.3
django-filebrowser==3.11.1
django-filebrowser-no-grappelli==3.7.8
django-froala-editor==2.9.5
django-grappelli==2.12.3
django-jet==1.0.8
django-js-asset==1.2.2
django-mptt==0.10.0
django-object-tools==2.0.0
django-polymorphic==2.0.3
django-tinymce4-lite==1.7.5
easy-thumbnails==2.6
google-api-python-client==1.4.1
httplib2==0.12.3
jsmin==2.2.2
oauth2client==4.1.3
pilkit==2.0
Pillow==6.0.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycodestyle==2.5.0
python-slugify==3.0.2
pytz==2019.1
rsa==4.0
six==1.12.0
slugify==0.0.1
sorl-thumbnail==12.5.0
sqlparse==0.3.0
text-unidecode==1.2
Unidecode==1.0.23
uritemplate==3.0.0
